+++ insurance-fund/programs/rlp/src/instructions/admin/mod.rs
@@ -4,9 +4,6 @@ pub use add_asset::*;
 pub mod initialize_rlp;
 pub use initialize_rlp::*;
 
-pub mod boost_rewards;
-pub use boost_rewards::*;
-
 pub mod manage_freeze;
 pub use manage_freeze::*;
 
@@ -16,5 +13,5 @@ pub use update_deposit_cap::*;
 pub mod initialize_lp;
 pub use initialize_lp::*;
 
-pub mod initialize_lp_lockup;
-pub use initialize_lp_lockup::*;
+pub mod initialize_lp_atas;
+pub use initialize_lp_atas::*;
