+++ insurance-fund/programs/rlp/src/instructions/user/mod.rs
@@ -6,9 +6,3 @@ pub use request_withdraw::*;
 
 pub mod withdraw;
 pub use withdraw::*;
-
-pub mod process_intent;
-pub use process_intent::*;
-
-pub mod deposit_asset;
-pub use deposit_asset::*;
