+++ insurance-fund/programs/rlp/src/instructions/user/restake.rs
@@ -1,261 +1,158 @@
 use anchor_lang::prelude::*;
-use anchor_spl::token;
+use anchor_spl::token::{Mint, TokenAccount, Token, transfer, Transfer};
+use spl_math::precise_number::PreciseNumber;
 use crate::errors::InsuranceFundError;
-use crate::events::RestakeEvent;
-use crate::helpers::calculate_receipts_on_mint;
-use crate::helpers::calculate_total_deposits;
-use crate::states::*;
+use crate::states::{Asset, LiquidityPool, Settings};
 use crate::constants::*;
-use anchor_spl::token::{
-    Mint,
-    Token,
-    TokenAccount
-};
 
 #[derive(AnchorDeserialize, AnchorSerialize)]
 pub struct RestakeArgs {
-    pub lockup_id: u64,
+    pub liquidity_pool_index: u8,
     pub amount: u64,
+    pub min_lp_tokens: u64,
 }
 
-#[inline(never)]
-pub fn restake(
-    ctx: Context<Restake>,
+pub fn restake<'info>(
+    ctx: Context<'_, 'info, 'info, 'info, Restake<'info>>,
     args: RestakeArgs
-) -> anchor_lang::prelude::Result<()> {
-
-    let RestakeArgs {
-        amount,
-        lockup_id: _
+) -> Result<()> {
+    let RestakeArgs { 
+        liquidity_pool_index: _, 
+        amount, 
+        min_lp_tokens 
     } = args;
 
-    let token_program = &ctx.accounts.token_program;
-    let user = &ctx.accounts.user;
-    let asset = &mut ctx.accounts.asset;
-    let asset_mint = &ctx.accounts.asset_mint;
-    let oracle = &ctx.accounts.oracle;
-    let lockup = &mut ctx.accounts.lockup;
-    let deposit = &mut ctx.accounts.deposit;
-    let lockup_cold_vault = &ctx.accounts.lockup_cold_vault;
-    let lockup_hot_vault = &ctx.accounts.lockup_hot_vault;
-    let receipt_token_mint = &ctx.accounts.receipt_token_mint;
-    let user_asset_ata = &ctx.accounts.user_asset_ata;
+    let signer = &ctx.accounts.signer;
     let settings = &ctx.accounts.settings;
-    let system_program = &ctx.accounts.system_program;
-    let deposit_receipt_token_account = &ctx.accounts.deposit_receipt_token_account;
+    let liquidity_pool = &ctx.accounts.liquidity_pool;
+    let lp_token = &ctx.accounts.lp_token;
+    let token_program = &ctx.accounts.token_program;
 
-    deposit.create_deposit_receipt_token_account(
-        &user,
-        &deposit_receipt_token_account,
-        ctx.bumps.deposit_receipt_token_account,
-        &deposit,
-        &receipt_token_mint,
-        &token_program,
-        &system_program
-    )?;
-
-    let total_deposits = calculate_total_deposits(
-        &lockup_cold_vault,
-        &lockup_hot_vault
-    )?;
-
-    match lockup.deposit_cap {
-        Some(cap) => {
-            require!(
-                total_deposits + amount <= cap,
-                InsuranceFundError::DepositCapOverflow
-            );
-        },
-        None => {}
-    }
-
-    let total_receipts_to_mint: u64 = calculate_receipts_on_mint(
-        &receipt_token_mint,
-        amount, 
-        total_deposits
-    )?;
-
-    lockup.mint_receipts(
-        total_receipts_to_mint, 
-        &lockup, 
-        &receipt_token_mint, 
-        &deposit_receipt_token_account, 
-        &token_program
-    )?;
+    require!(
+        amount > 0,
+        crate::errors::InsuranceFundError::InvalidInput
+    );
 
     let clock = Clock::get()?;
-    let unix_ts = clock.unix_timestamp as u64;
 
-    deposit.index = lockup.deposits;
-    deposit.lockup = lockup.key();
-    deposit.user = user.key();
-    deposit.unlock_ts = unix_ts + lockup.duration;
-    deposit.bump = ctx.bumps.deposit;
+    let total_pool_value_before = liquidity_pool.calculate_total_pool_value(
+        &ctx.remaining_accounts,
+        liquidity_pool,
+        settings,
+        &clock
+    )?;
 
-    let receipt_exchange_rate_bps = lockup.receipt_to_reward_exchange_rate_bps_accumulator;
-    deposit.initial_receipt_exchange_rate_bps = receipt_exchange_rate_bps;
-    
-    let oracle_price = asset
-        .get_price(oracle, &clock)?;
-
-    // Multiply oracle price by amount of tokens, 
-    // then divide by token decimals to get value of 1 full token instead of `lamports`.
-    let usd_value_with_decimals = oracle_price
-        .mul(amount)?;
-
-    let usd_value: u64 = usd_value_with_decimals
-        .checked_div(
-            10_u128
-            .checked_pow(asset_mint.decimals as u32)
-            .ok_or(InsuranceFundError::MathOverflow)?
-        )
-        .ok_or(InsuranceFundError::MathOverflow)?
-        .try_into()
-        .map_err(|_| InsuranceFundError::MathOverflow)?;
-
-    deposit.initial_usd_value = usd_value;
-
-    lockup.deposit(
-        amount, 
-        user, 
-        user_asset_ata, 
-        settings, 
-        lockup_hot_vault, 
-        lockup_cold_vault, 
+    liquidity_pool.deposit(
+        signer,
+        amount,
+        &ctx.accounts.user_asset_account,
+        &ctx.accounts.pool_asset_account,
         token_program
     )?;
 
-    lockup.deposits += 1;
+    let asset = &ctx.accounts.asset;
+    let oracle = &ctx.accounts.oracle;
+    let deposit_asset_price = asset.get_price(oracle, &clock)?;
 
-    asset.increase_tvl(amount)?;
-    asset.add_deposit()?;
+    let deposit_value = PreciseNumber::new(
+        deposit_asset_price.mul(amount)?
+    ).ok_or(InsuranceFundError::MathOverflow)?;
 
-    emit!(RestakeEvent {
-        from: user.key(),
-        asset: asset.key(),
-        amount,
-        lockup_ts: deposit.unlock_ts
-    });
+    let lp_tokens_to_mint = liquidity_pool
+        .calculate_lp_tokens_on_deposit(
+            lp_token,
+            total_pool_value_before,
+            deposit_value
+        )?;
+
+    require!(
+        min_lp_tokens <= lp_tokens_to_mint,
+        InsuranceFundError::SlippageExceeded
+    );
+
+    liquidity_pool.mint_lp_token(
+        lp_tokens_to_mint,
+        liquidity_pool,
+        lp_token,
+        &ctx.accounts.user_lp_account,
+        token_program
+    )?;
 
     Ok(())
 }
 
 #[derive(Accounts)]
-#[instruction(
-    args: RestakeArgs
-)]
+#[instruction(args: RestakeArgs)]
 pub struct Restake<'info> {
-    #[account(
-        mut
-    )]
-    pub user: Signer<'info>,
+    #[account(mut)]
+    pub signer: Signer<'info>,
 
     #[account(
         mut,
         seeds = [
-            SETTINGS_SEED.as_bytes()
+            SETTINGS_SEED.as_bytes(),
         ],
         bump,
-        constraint = !settings.frozen @ InsuranceFundError::Frozen
+        constraint = !settings.frozen @ InsuranceFundError::Frozen,
     )]
-    pub settings: Box<Account<'info, Settings>>,
+    pub settings: Account<'info, Settings>,
 
     #[account(
-        mut,
         seeds = [
-            LOCKUP_SEED.as_bytes(),
-            &args.lockup_id.to_le_bytes()
+            LIQUIDITY_POOL_SEED.as_bytes(),
+            &args.liquidity_pool_index.to_le_bytes()
         ],
         bump,
-        constraint = lockup.min_deposit <= args.amount @ InsuranceFundError::DepositTooLow,
+        constraint = liquidity_pool.index == args.liquidity_pool_index,
     )]
-    pub lockup: Box<Account<'info, Lockup>>,
+    pub liquidity_pool: Account<'info, LiquidityPool>,
 
     #[account(
         mut,
-        address = lockup.receipt_mint
+        address = liquidity_pool.lp_token
     )]
-    pub receipt_token_mint: Account<'info, Mint>,
-
-    #[account(
-        init,
-        payer = user,
-        seeds = [
-            DEPOSIT_SEED.as_bytes(),
-            lockup.key().as_ref(),
-            &lockup.deposits.to_le_bytes()
-        ],
-        bump,
-        space = 8+ Deposit::INIT_SPACE
-    )]
-    pub deposit: Account<'info, Deposit>,
-
-    /// CHECK: initializing this manually
-    #[account(
-        mut,
-        seeds = [
-            DEPOSIT_RECEIPT_VAULT_SEED.as_bytes(),
-            deposit.key().as_ref(),
-            receipt_token_mint.key().as_ref(),
-        ],
-        bump
-    )]
-    pub deposit_receipt_token_account: AccountInfo<'info>,
+    pub lp_token: Box<Account<'info, Mint>>,
 
     #[account(
         mut,
+        associated_token::mint = lp_token,
+        associated_token::authority = signer,
+    )]
+    pub user_lp_account: Box<Account<'info, TokenAccount>>,
+
+    #[account(
         seeds = [
             ASSET_SEED.as_bytes(),
             &asset_mint.key().to_bytes()
         ],
         bump,
-        constraint = asset.mint == asset_mint.key()
     )]
     pub asset: Account<'info, Asset>,
 
     #[account(
         mut,
-        address = lockup.asset_mint,
+        address = asset.mint
     )]
-    pub asset_mint: Account<'info, Mint>,
+    pub asset_mint: Box<Account<'info, Mint>>,
 
     #[account(
         mut,
-        seeds = [
-            HOT_VAULT_SEED.as_bytes(),
-            lockup.key().as_ref(),
-            asset_mint.key().as_ref(),
-        ],
-        bump,
+        token::mint = asset_mint,
+        token::authority = signer,
     )]
-    pub lockup_hot_vault: Account<'info, anchor_spl::token::TokenAccount>,
-
-    /// CHECK: DEBUG
-    #[account(
-        mut,
-        seeds = [
-            COLD_VAULT_SEED.as_bytes(),
-            lockup.key().as_ref(),
-            asset_mint.key().as_ref(),
-        ],
-        bump,
-    )]
-    pub lockup_cold_vault: Account<'info, anchor_spl::token::TokenAccount>,
+    pub user_asset_account: Box<Account<'info, TokenAccount>>,
 
     #[account(
         mut,
-        associated_token::authority = user,
         associated_token::mint = asset_mint,
-        constraint = user_asset_ata.amount >= args.amount @ InsuranceFundError::NotEnoughFunds,
+        associated_token::authority = liquidity_pool,
     )]
-    pub user_asset_ata: Account<'info, TokenAccount>,
+    pub pool_asset_account: Box<Account<'info, TokenAccount>>,
 
-    /// CHECK: Directly checking the address + checking type later.
     #[account(
-        mut,
-        constraint = asset.oracle.key().eq(&oracle.key())
+        constraint = asset.oracle.key() == &oracle.key()
     )]
-    pub oracle: AccountInfo<'info>,
+    pub oracle: UncheckedAccount<'info>,
 
     #[account()]
     pub token_program: Program<'info, Token>,
