+++ insurance-fund/programs/rlp/src/states/lp/liquidity_pool.rs
@@ -2,13 +2,17 @@ use anchor_lang::prelude::*;
 use anchor_spl::token::{
     mint_to, transfer, Mint, MintTo, Token, TokenAccount, Transfer
 };
-use crate::{constants::LIQUIDITY_POOL_SEED, errors::InsuranceFundError, helpers::OraclePrice};
+use crate::constants::*;
+use crate::helpers::OraclePrice;
+use spl_math::precise_number::PreciseNumber;
+use crate::errors::InsuranceFundError;
+use crate::states::*;
 
 #[derive(InitSpace)]
 #[account]
 pub struct LiquidityPool {
     pub bump: u8,
-    pub index: u64,
+    pub index: u8,
     pub lp_token: Pubkey
 }
 
@@ -38,40 +42,121 @@ impl LiquidityPool {
         Ok(())
     }
 
-    /// DEPRECATED: This method only considers a single asset. Use the new deposit_asset instruction
-    /// that calculates LP tokens based on the total USD value of the entire liquidity pool.
-    pub fn compute_lp_tokens_on_deposit<'info>(
+    pub fn calculate_total_pool_value<'info>(
         &self,
-        lp_token_supply: u64,
-        asset_liquidity: u64,
-        asset_deposit: u64,
-        asset_price: OraclePrice,
-    ) -> Result<u64> {
-        let deposit_value = asset_price
-            .mul(asset_deposit)?;
-
-        if lp_token_supply == 0 {
-            let result: u64 = deposit_value
-                .try_into()
-                .map_err(|_| InsuranceFundError::MathOverflow)?;
-
-            return Ok(result);
+        remaining_accounts: &'info [AccountInfo<'info>],
+        liquidity_pool: &'info Account<'info, LiquidityPool>,
+        settings: &'info Account<'info, Settings>,
+        clock: &Clock,
+    ) -> Result<PreciseNumber> {
+        let mut total_pool_value = PreciseNumber::new(0)
+            .ok_or(crate::errors::InsuranceFundError::MathOverflow)?;
+        
+        require!(
+            remaining_accounts.len() == settings.assets as usize * 3,
+            crate::errors::InsuranceFundError::InvalidInput
+        );
+        
+        let mut i = 0;
+        while i < remaining_accounts.len() {
+            let token_account_info = &remaining_accounts[i];
+            
+            require!(
+                token_account_info.owner == &anchor_spl::token::ID,
+                crate::errors::InsuranceFundError::InvalidInput
+            );
+    
+            let token_account = TokenAccount::try_deserialize(&mut token_account_info.try_borrow_mut_data()?.as_ref())
+                .map_err(|_| crate::errors::InsuranceFundError::InvalidInput)?;
+    
+            require!(
+                token_account.owner == liquidity_pool.key(),
+                crate::errors::InsuranceFundError::InvalidInput
+            );
+    
+            let asset_info = &remaining_accounts[i + 1];
+            
+            require!(
+                asset_info.owner == &crate::ID,
+                crate::errors::InsuranceFundError::InvalidInput
+            );
+            
+            let asset = Asset::try_deserialize(&mut asset_info.try_borrow_mut_data()?.as_ref())
+                .map_err(|_| crate::errors::InsuranceFundError::InvalidInput)?;
+    
+            require!(
+                asset.mint == token_account.mint,
+                crate::errors::InsuranceFundError::InvalidInput
+            );
+            
+            let (expected_asset_pda, _) = Pubkey::find_program_address(
+                &[
+                    crate::constants::ASSET_SEED.as_bytes(),
+                    &asset.mint.to_bytes()
+                ],
+                &crate::ID
+            );
+    
+            require!(
+                asset_info.key() == expected_asset_pda,
+                crate::errors::InsuranceFundError::InvalidInput
+            );
+    
+            let oracle_info = &remaining_accounts[i + 2];
+    
+            require!(
+                oracle_info.key() == *asset.oracle.key(),
+                crate::errors::InsuranceFundError::InvalidInput
+            );
+    
+            let asset_price = asset.get_price(oracle_info, clock)
+                .map_err(|_| crate::errors::InsuranceFundError::InvalidInput)?;
+    
+            let token_balance = token_account.amount;
+            if token_balance > 0 {
+                let token_value_precise = PreciseNumber::new(asset_price.mul(token_balance)?)
+                    .ok_or(crate::errors::InsuranceFundError::MathOverflow)?;
+                total_pool_value = total_pool_value.checked_add(&token_value_precise)
+                    .ok_or(crate::errors::InsuranceFundError::MathOverflow)?;
+            }
+    
+            i += 3;
         }
+    
+        Ok(total_pool_value)
+    } 
 
-        let asset_liquidity_value = asset_price
-            .mul(asset_liquidity)?;
+    pub fn calculate_lp_tokens_on_deposit<'info>(
+        &self,
+        lp_token: &'info Account<'info, Mint>,
+        total_pool_value: PreciseNumber,
+        deposit_value: PreciseNumber,
+    ) -> Result<u64> {
+        let lp_tokens_to_mint = if lp_token.supply == 0 {
+            deposit_value.floor()
+                .ok_or(crate::errors::InsuranceFundError::MathOverflow)?
+                .to_imprecise()
+                .ok_or(crate::errors::InsuranceFundError::MathOverflow)?
+                .try_into()
+                .map_err(|_| crate::errors::InsuranceFundError::MathOverflow)?
+        } else {
+            let lp_supply_precise = PreciseNumber::new(lp_token.supply as u128)
+                .ok_or(crate::errors::InsuranceFundError::MathOverflow)?;
+            let deposit_ratio = deposit_value
+                .checked_mul(&lp_supply_precise)
+                .ok_or(crate::errors::InsuranceFundError::MathOverflow)?
+                .checked_div(&total_pool_value)
+                .ok_or(crate::errors::InsuranceFundError::MathOverflow)?;
+            
+            deposit_ratio.floor()
+                .ok_or(crate::errors::InsuranceFundError::MathOverflow)?
+                .to_imprecise()
+                .ok_or(crate::errors::InsuranceFundError::MathOverflow)?
+                .try_into()
+                .map_err(|_| crate::errors::InsuranceFundError::MathOverflow)?
+        };
 
-        let value_per_lp_token = asset_liquidity_value
-            .checked_div(lp_token_supply.into())
-            .ok_or(InsuranceFundError::MathOverflow)?;
-
-        let lp_token_to_mint: u64 = deposit_value
-            .checked_div(value_per_lp_token)
-            .ok_or(InsuranceFundError::MathOverflow)?
-            .try_into()
-            .map_err(|_| InsuranceFundError::MathOverflow)?;
-
-        Ok(lp_token_to_mint)
+        Ok(lp_tokens_to_mint)
     }
 
     pub fn mint_lp_token<'info>(
