+++ insurance-fund/programs/rlp/src/states/lp/mod.rs
@@ -1,5 +1,2 @@
 pub mod liquidity_pool;
 pub use liquidity_pool::*;
-
-pub mod lp_lockup;
-pub use lp_lockup::*;
